+++ msbuild/Xamarin.Mac.Tasks/Xamarin.Mac.Common.targets
@@ -457,6 +457,7 @@ Copyright (C) 2014 Xamarin. All rights reserved.
 			<Output TaskParameter="SdkDevPath" PropertyName="_SdkDevPath" />
 			<Output TaskParameter="SdkUsrPath" PropertyName="_SdkUsrPath" />
 			<Output TaskParameter="IsXcode8" PropertyName="_IsXcode8" />
+			<Output TaskParameter="XamarinSdkRoot" PropertyName="_XamarinSdkRoot" />
 		</DetectSdkLocations>
 	</Target>
 
@@ -531,11 +532,10 @@ Copyright (C) 2014 Xamarin. All rights reserved.
 		<Mmp
 			Condition="'$(IsMacEnabled)' == 'true'"
 			SessionId="$(BuildSessionId)"
-			FrameworkRoot="$(XamarinMacFrameworkRoot)"
 			AppBundleDir="$(AppBundleDir)"
 			OutputPath="$(OutputPath)"
 			ApplicationName="$(_AppBundleName)"
-			ApplicationAssembly="$(OutputPath)$(TargetName)$(TargetExt)"
+			MainAssembly="$(OutputPath)$(TargetName)$(TargetExt)"
 			TargetFrameworkMoniker="$(_ComputedTargetFrameworkMoniker)"
 			UseXamMacFullFramework="$(UseXamMacFullFramework)"
 			Architecture="$(XamMacArch)"
@@ -545,9 +545,9 @@ Copyright (C) 2014 Xamarin. All rights reserved.
 			HttpClientHandler="$(HttpClientHandler)"
 			I18n="$(I18n)"
 			Profiling="$(Profiling)"
-			ExtraArguments="$(MonoBundlingExtraArgs)"
+			ExtraArgs="$(MonoBundlingExtraArgs)"
 			NativeReferences="@(NativeReference)"
-			ExplicitReferences="@(ReferencePath);@(ReferenceCopyLocalAssemblyPaths)"
+			References="@(ReferencePath);@(ReferenceCopyLocalAssemblyPaths)"
 			ResponseFilePath="$(IntermediateOutputPath)response-file.rsp"
 			SdkRoot="$(_SdkDevPath)"
 			IntermediateOutputPath="$(IntermediateOutputPath)mmp-cache"
@@ -560,6 +560,7 @@ Copyright (C) 2014 Xamarin. All rights reserved.
 			HybridAOT="$(HybridAOT)"
 			ExplicitAotAssemblies="$(ExplicitAotAssemblies)"
 			StandardOutputImportance="High"
+			XamarinSdkRoot="$(_XamarinSdkRoot)"
 			>
 			<Output TaskParameter="NativeLibraries" ItemName="_NativeLibrary" />
 		</Mmp>
